+++ src/documents/fields/SessionLocation.ts
@@ -4,5 +4,6 @@ export const SessionLocationSchema = () => (
   new fields.SchemaField({
     uuid: new fields.DocumentUUIDField({ required: true, nullable: false }),
     delivered: new fields.BooleanField({ required: true, nullable: false }),
+    notes: new fields.StringField({ required: true, nullable: false, initial: '' }),
   }, { required: true, nullable: false} )
 );