+++ src/documents/fields/SessionMonster.ts
@@ -5,5 +5,6 @@ export const SessionMonsterSchema = () => (
     uuid: new fields.DocumentUUIDField({ required: true, nullable: false }),
     delivered: new fields.BooleanField({ required: true, nullable: false }),
     number: new fields.NumberField({ required: true, nullable: false, integer: true }),
+    notes: new fields.StringField({ required: true, nullable: false, initial: '' }),
   }, { required: true, nullable: false} )
 );